Add ill tile modifiers to config

+++ Sources/logger/inf_logger.cpp
@@ -6,6 +6,6 @@ inf_logger::inf_logger(std::string filename) : m_log(filename) {}
 
 void inf_logger::log(simtime_t time, std::size_t x, std::size_t y,
                      tile_type tt) {
-  m_log << time << " " << x << " " << y << " " << tt << "\n";
+  m_log << time << " " << x << " " << y << " " << static_cast<long>(tt) << "\n";
 }
 } // namespace sprsim
